[GREEN] add authorization to controller

+++ src/main/java/hoomgroom/transaction/transaksi/controller/TransaksiController.java
@@ -1,12 +1,17 @@
 package hoomgroom.transaction.transaksi.controller;
 
+import hoomgroom.transaction.Auth.model.Role;
+import hoomgroom.transaction.Auth.model.User;
+import hoomgroom.transaction.Auth.service.JwtService;
 import hoomgroom.transaction.transaksi.dto.RequestTransaksiData;
 import hoomgroom.transaction.transaksi.dto.TransaksiData;
 import hoomgroom.transaction.transaksi.service.TransaksiService;
+import lombok.NonNull;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import jakarta.servlet.http.HttpServletRequest;
 
 import java.util.List;
 import java.util.UUID;
@@ -15,21 +20,50 @@ import java.util.concurrent.CompletableFuture;
 @RestController
 @RequestMapping("/api/transaksi")
 public class TransaksiController {
+    @Autowired
+    JwtService jwtService;
 
     @Autowired
     private TransaksiService transaksiService;
 
+    private static final String JWT_HEADER = "Authorization";
+    private static final String JWT_TOKEN_PREFIX = "Bearer";
+
     @RequestMapping(value = "/view", method = RequestMethod.GET)
-    public CompletableFuture<ResponseEntity<List<TransaksiData>>> getAllTransaksi() {
+    public CompletableFuture<ResponseEntity> getAllTransaksi(@NonNull HttpServletRequest request) {
+        final String authHeader = request.getHeader(JWT_HEADER);
+        if (authHeader == null) {
+            return CompletableFuture.supplyAsync(() -> ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized"));
+        }
+        String jwtToken = authHeader.substring(JWT_TOKEN_PREFIX.length());
+        User userDetails = jwtService.extractUser(jwtToken);
+
+        String username = userDetails.getUsername();
+        Role role = userDetails.getRole();
+        if (role == Role.USER) {
+            return CompletableFuture.supplyAsync(() -> ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized"));
+        }
         return CompletableFuture.supplyAsync(() -> ResponseEntity.ok(transaksiService.findAll()));
     }
 
     @RequestMapping(value = "/view/{id}", method = RequestMethod.GET)
-    public CompletableFuture<ResponseEntity> getTransaksiById(@PathVariable UUID id) {
+    public CompletableFuture<ResponseEntity> getTransaksiById(@NonNull HttpServletRequest request, @PathVariable UUID id) {
+        final String authHeader = request.getHeader(JWT_HEADER);
+        if (authHeader == null) {
+            return CompletableFuture.supplyAsync(() -> ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized"));
+        }
+        String jwtToken = authHeader.substring(JWT_TOKEN_PREFIX.length());
+        User userDetails = jwtService.extractUser(jwtToken);
+
+        String username = userDetails.getUsername();
+        Role role = userDetails.getRole();
         return CompletableFuture.supplyAsync(() -> {
             try {
                 TransaksiData transaksiData = transaksiService.findById(id);
-                return ResponseEntity.ok().body(transaksiData);
+                if (transaksiData.getUsername().equals(username) || role == Role.ADMIN) {
+                    return ResponseEntity.ok().body(transaksiData);
+                }
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized");
             } catch (Exception e) {
                 return ResponseEntity.badRequest().body(HttpStatus.BAD_REQUEST);
             }
@@ -37,7 +71,16 @@ public class TransaksiController {
     }
 
     @RequestMapping(value = "/create", method = RequestMethod.POST)
-    public CompletableFuture<ResponseEntity> createTransaksi(@RequestBody RequestTransaksiData transaksiData) {
+    public CompletableFuture<ResponseEntity> createTransaksi(@NonNull HttpServletRequest request, @RequestBody RequestTransaksiData transaksiData) {
+        final String authHeader = request.getHeader(JWT_HEADER);
+        if (authHeader == null) {
+            return CompletableFuture.supplyAsync(() -> ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized"));
+        }
+        String jwtToken = authHeader.substring(JWT_TOKEN_PREFIX.length());
+        User userDetails = jwtService.extractUser(jwtToken);
+
+        String username = userDetails.getUsername();
+        transaksiData.setUsername(username);
         return CompletableFuture.supplyAsync(() -> {
             try {
                 TransaksiData transaksiData1 = transaksiService.createTransaksi(transaksiData);
@@ -49,10 +92,23 @@ public class TransaksiController {
     }
 
     @RequestMapping(value = "/delete/{id}", method = RequestMethod.DELETE)
-    public CompletableFuture<ResponseEntity<Void>> deleteTransaksi(@PathVariable UUID id) {
+    public CompletableFuture<ResponseEntity> deleteTransaksi(@NonNull HttpServletRequest request, @PathVariable UUID id) {
+        final String authHeader = request.getHeader(JWT_HEADER);
+        if (authHeader == null) {
+            return CompletableFuture.supplyAsync(() -> ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized"));
+        }
+        String jwtToken = authHeader.substring(JWT_TOKEN_PREFIX.length());
+        User userDetails = jwtService.extractUser(jwtToken);
+
+        String username = userDetails.getUsername();
+        Role role = userDetails.getRole();
         return CompletableFuture.supplyAsync(() -> {
             try {
-                transaksiService.delete(id);
+                TransaksiData transaksiData = transaksiService.findById(id);
+                if (transaksiData.getUsername().equals(username) || role == Role.ADMIN) {
+                    transaksiService.delete(id);
+                    return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized");
+                }
                 return new ResponseEntity<>(HttpStatus.NO_CONTENT);
             } catch (RuntimeException e) {
                 return new ResponseEntity<>(HttpStatus.NOT_FOUND);
@@ -61,11 +117,18 @@ public class TransaksiController {
     }
 
     @RequestMapping(value = "/filter", method = RequestMethod.GET)
-    public CompletableFuture<ResponseEntity<List<TransaksiData>>> filterTransaksi(
-            @RequestParam(required = false) String username,
+    public CompletableFuture<ResponseEntity> filterTransaksi(
+            @NonNull HttpServletRequest request,
             @RequestParam(required = false, defaultValue = "false") boolean time,
             @RequestParam(required = false, defaultValue = "true") boolean isAscending) {
+        final String authHeader = request.getHeader(JWT_HEADER);
+        if (authHeader == null) {
+            return CompletableFuture.supplyAsync(() -> ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized"));
+        }
+        String jwtToken = authHeader.substring(JWT_TOKEN_PREFIX.length());
+        User userDetails = jwtService.extractUser(jwtToken);
 
+        String username = userDetails.getUsername();
         return CompletableFuture.supplyAsync(() -> {
             try {
                 List<TransaksiData> filteredTransaksi = transaksiService.findByFilter(username, time, isAscending);
